Improve developer ergonomics
#17, #18, #19

Several members of our team are now using nvi for their daily
development workflow, which has brought some less than ergonomic traits
of nvi to the surface. See the referenced issues for a more thorough
explanation.

While these issues are mostly irrelevant in a CI setting, which is the
intended environment of nvi, there is no harm in fixing them. This
commit addresses all three issues:

1. There should be no more warnings about null bytes in input.
2. If creating the node symlink fails, nvi doesn't attempt to create the
   npm symlink.
3. The node executable is now a symlink (like the npm executable), which
   means nvi can change it even when the linked node executable is
   running.

The fix of (3) in particular should greatly improve the experience of
using nvi in development workflows.