Link npx executable alongside npm and node
Since npm@5.2.0 we have been getting npx installed as well as node and
npm. It is about time we also include this when installing node with
nvi, since we have taken to using it for local development as well as
the CI setup it was originally intended for.

This does not fix the issue with globally installed packages mentioned
in #23, but it does help circumvent it, since you can run things through
npx for most of these cases.