Support alternative origin
The official nodejs.org domain hasn't been a very reliable origin for us
for downloading binaries from lately. Often it works perfectly fine, but
occasionally it seems to just not send any bytes at all after
establishing a connection.

Most likely nodejs.org are applying rate limits, which would make good
sense from a cost and abuse perspective. However, this makes it very
unreliable for CI tools like `nvi`.

To provide a work around for this, we're adding a flag to change to a
different origin. Most likely this alternative origin will be a caching
proxy, which is certainly what we'll be using it for.